tidy: Rename tools/tools.rs to tools/specs.rs
